requirements.txt
-e .
tox
pytest-django==4.5.2
